Add <CustomTheme type>
