+++ src/components/common/Callout.tsx
@@ -5,13 +5,7 @@ import { Theme, ThemeProvider } from '@rmwc/theme';
 import { Typography } from '@rmwc/typography';
 import React, { ReactNode } from 'react';
 
-import {
-  cautionTheme,
-  errorTheme,
-  noteTheme,
-  successTheme,
-  tipTheme,
-} from '../../themes';
+import { CustomTheme, CustomThemeType, Type } from '../../themes';
 
 export interface CalloutProps {
   /** Custom icon to override the default */
@@ -19,19 +13,10 @@ export interface CalloutProps {
   /** Show a lighter version of the callout */
   aside?: true;
   /** The color scheme of the callout */
-  type?: 'note' | 'success' | 'warning' | 'caution' | 'tip';
+  type?: CustomThemeType;
 
   /** Action buttons */
   actions?: ReactNode;
-}
-
-// TODO: Fill in tip (based on)
-export enum Type {
-  CAUTION = 'caution',
-  NOTE = 'note',
-  SUCCESS = 'success',
-  TIP = 'tip',
-  WARNING = 'warning',
 }
 
 const DEFAULT_ICON_MAP: Record<string, string> = {
@@ -40,14 +25,6 @@ const DEFAULT_ICON_MAP: Record<string, string> = {
   [Type.SUCCESS]: 'check_circle',
   [Type.TIP]: 'star',
   [Type.WARNING]: 'warning',
-};
-
-const THEME_MAP: Record<string, Record<string, string>> = {
-  [Type.CAUTION]: cautionTheme,
-  [Type.NOTE]: noteTheme,
-  [Type.SUCCESS]: successTheme,
-  [Type.TIP]: tipTheme,
-  [Type.WARNING]: errorTheme,
 };
 
 export const Callout: React.FC<CalloutProps> = ({
@@ -62,7 +39,7 @@ export const Callout: React.FC<CalloutProps> = ({
   const classes = `Callout Callout-${type}${asideClass}`;
 
   return (
-    <ThemeProvider options={THEME_MAP[type]} wrap>
+    <CustomTheme use={type} wrap>
       <Theme
         className={classes}
         use={aside ? [] : ['onPrimary', 'primaryBg']}
@@ -85,6 +62,6 @@ export const Callout: React.FC<CalloutProps> = ({
           <div className="Callout-actions">{actions}</div>
         </ThemeProvider>
       </Theme>
-    </ThemeProvider>
+    </CustomTheme>
   );
 };
